pkg/workspace/graphql: rename files
--- sturdy ---
user_id: cc33c18b-27e5-470d-8589-f9391ca37590
view_id: 2c064d23-2410-4e86-939f-91c56dae2fae
workspace_id: 855151f1-f5c7-41f7-b2e0-833fb3f35550
